# api-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agentic AI Agricultural Crop Planning & Precision Farming System",
    "summary": "Agentic multi-agent crop planning and precision farming API.",
    "description": "\nMulti-agent decision support for crop planning and precision farming.\n\nThe workflow is orchestrated with **LangGraph** over twelve specialised agents\n(farm profile, soil & nutrient, weather & climate, agricultural knowledge RAG,\ncrop suitability, ML forecasting, irrigation, crop-risk review, activity\nplanning and advisory). Every agent emits structured evidence, and every\nphysical action (irrigation in particular) requires **human approval** before it\ncan leave the `planned` state.\n",
    "contact": {
      "name": "Precision Farming Platform"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/suitability/assess": {
      "post": {
        "tags": [
          "crop suitability"
        ],
        "summary": "Assess crop suitability for a field",
        "operationId": "assess_api_v1_suitability_assess_post",
        "parameters": [
          {
            "name": "field_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuitabilityRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuitabilityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/suitability/fields/{field_id}": {
      "get": {
        "tags": [
          "crop suitability"
        ],
        "summary": "History",
        "operationId": "suitability_history_api_v1_suitability_fields__field_id__get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SuitabilityOut"
                  },
                  "title": "Response Suitability History Api V1 Suitability Fields  Field Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/irrigation/assess": {
      "post": {
        "tags": [
          "irrigation"
        ],
        "summary": "Irrigation decision support (proposal only, never auto-applied)",
        "operationId": "assess_irrigation_api_v1_irrigation_assess_post",
        "parameters": [
          {
            "name": "field_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IrrigationAssessRequest",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IrrigationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/irrigation/fields/{field_id}": {
      "get": {
        "tags": [
          "irrigation"
        ],
        "summary": "History",
        "operationId": "irrigation_history_api_v1_irrigation_fields__field_id__get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IrrigationOut"
                  },
                  "title": "Response Irrigation History Api V1 Irrigation Fields  Field Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/fields/{field_id}": {
      "get": {
        "tags": [
          "crop risk"
        ],
        "summary": "Scan environmental risk",
        "operationId": "scan_risk_api_v1_risk_fields__field_id__get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          },
          {
            "name": "crop",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Crop"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/fields/{field_id}/history": {
      "get": {
        "tags": [
          "crop risk"
        ],
        "summary": "Persisted findings",
        "operationId": "risk_history_api_v1_risk_fields__field_id__history_get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RiskOut"
                  },
                  "title": "Response Risk History Api V1 Risk Fields  Field Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ml/predict": {
      "post": {
        "tags": [
          "machine learning"
        ],
        "summary": "Run both models against a field's current conditions",
        "description": "Score a field with both trained models.\n\nThe body is optional: a bare ``POST /ml/predict?field_id=N`` scores the field\non its own stored context (sensor telemetry, soil sample and live forecast),\nand any body field simply overrides that context.\n\nFeature assembly is delegated to :mod:`app.services.ml_service` so that this\nendpoint and the ``ml_forecast_agent`` are guaranteed to feed the models the\nsame, real values.  Building the vector from the request body alone would\nsilently median-impute every soil and weather feature and report the result\nas if it had been measured.",
        "operationId": "predict_api_v1_ml_predict_post",
        "parameters": [
          {
            "name": "field_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MLPredictRequest",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MLPredictionOut"
                  },
                  "title": "Response Predict Api V1 Ml Predict Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ml/fields/{field_id}/predictions": {
      "get": {
        "tags": [
          "machine learning"
        ],
        "summary": "Prediction history",
        "operationId": "prediction_history_api_v1_ml_fields__field_id__predictions_get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MLPredictionOut"
                  },
                  "title": "Response Prediction History Api V1 Ml Fields  Field Id  Predictions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IrrigationAssessRequest": {
        "properties": {
          "crop": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crop"
          },
          "crop_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crop Stage"
          },
          "soil_moisture_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soil Moisture Percent"
          }
        },
        "type": "object",
        "title": "IrrigationAssessRequest"
      },
...
      "SuitabilityRequest": {
        "properties": {
          "crop": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crop",
            "description": "Defaults to the field's proposed crop"
          },
          "soil_observation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soil Observation Id"
          },
          "include_evidence": {
            "type": "boolean",
            "title": "Include Evidence",
            "default": true
          }
        },
        "type": "object",
        "title": "SuitabilityRequest"
      },
...
    }
  }
}